Notes
The create function accepts two arguments. The first is: it’s a block containing name-value pairs that provide values for fields for the new account. If our Quick Account form had more fields, you could simply add appropriate name-value pairs to this block to create the account record with the additional field values. The second argument is a callback function. This function is called with the results of the Remote Objects operation once it completes.